Add Splashtop Business recipe
Splashtop made what could be an easy server-side render a very difficult
JavaScript client-side render. It's probably to prevent scraping. It
would be more effective for them to actually update their Sparkle feed
more than once per decade.

The downloadx.php script seems to expect r to be a random floating-point
number. But since we aren't able to do that in autopkg easily, we'll
just fake a super-old version of Firefox and hope it doesn't trigger any
alarm bells.